+++ geom/type_geometry.go
@@ -908,3 +908,33 @@ func (g Geometry) Summary() string {
 func (g Geometry) String() string {
 	return g.Summary()
 }
+
+// Simplify returns a simplified version of the geometry using the
+// Ramer-Douglas-Peucker algorithm. Sometimes a simplified geometry can become
+// invalid, in which case an error is returned rather than attempting to fix
+// the geometry. Validation of the result can be skipped by making use of the
+// geometry constructor options.
+func (g Geometry) Simplify(threshold float64, opts ...ConstructorOption) (Geometry, error) {
+	switch g.gtype {
+	case TypeGeometryCollection:
+		c, err := g.MustAsGeometryCollection().Simplify(threshold, opts...)
+		return c.AsGeometry(), err
+	case TypePoint:
+		return g, nil
+	case TypeLineString:
+		c := g.MustAsLineString().Simplify(threshold)
+		return c.AsGeometry(), nil
+	case TypePolygon:
+		c, err := g.MustAsPolygon().Simplify(threshold, opts...)
+		return c.AsGeometry(), err
+	case TypeMultiPoint:
+		return g, nil
+	case TypeMultiLineString:
+		return g.MustAsMultiLineString().Simplify(threshold).AsGeometry(), nil
+	case TypeMultiPolygon:
+		c, err := g.MustAsMultiPolygon().Simplify(threshold, opts...)
+		return c.AsGeometry(), err
+	default:
+		panic("unknown type: " + g.Type().String())
+	}
+}