+++ geom/type_multi_line_string.go
@@ -482,3 +482,20 @@ func (m MultiLineString) Summary() string {
 func (m MultiLineString) String() string {
 	return m.Summary()
 }
+
+// Simplify returns a simplified version of the MultiLineString by using the
+// Ramer-Douglas-Peucker algorithm on each of the child LineStrings. If the
+// Ramer-Douglas-Peucker were to create an invalid child LineString (i.e. one
+// having only a single distinct point), then it is omitted in the output.
+// Empty child LineStrings are also omitted from the output.
+func (m MultiLineString) Simplify(threshold float64) MultiLineString {
+	n := m.NumLineStrings()
+	lss := make([]LineString, 0, n)
+	for i := 0; i < n; i++ {
+		ls := m.LineStringN(i).Simplify(threshold)
+		if !ls.IsEmpty() {
+			lss = append(lss, ls)
+		}
+	}
+	return NewMultiLineString(lss)
+}